Create SSR module
